Add nanomorph-component-id data attribute (#112)
This allows elements to opt out of being morphed with elements from a
different source, such as elements from component A vs component B.

This matters for component systems that "own" their DOM subtree.
Nanocomponent does not own its tree, and doing `this.element` looks
it up in the document. `this.element` may refer to different real
nodes at different times. But other methods are possible, for example:

```js
class BaseComponent {
  constructor (name) {
    this.name = name
    this.element = this.createElement()
    this.element.dataset.nanomorphComponentId = this.name
  }
  render () {
    const newElement = this.createElement()
    newElement.dataset.nanomorphComponentId = this.name
    const morphedElement = morph(this.element, newElement)
    assert(this.element === morphedElement, 'morphing did not replace the element')

    // don't need a proxy node, because `this.element` is always the same element
    return this.element
  }
}

class ComponentA extends BaseComponent {
  constructor () { super('A') }
  createElement () {
    return html`
      <div>
        component of type A
      </div>
    `
  }
}

class ComponentB extends BaseComponent {
  constructor () { super('B') }
  createElement () {
    return html`
      <div>
        component of type B
      </div>
    `
  }
}
```

Here, `BaseComponent` has an `element` property that always contains the
most up to date tree. nanomorph should never touch nodes from this
component, even if it's unmounted.

`isSameNode` can be used to prevent nanomorph from entering a subtree,
but when a component A is replaced by a component B, `isSameNode()`
will returns `false`, and nanomorph will attempt to morph the tree from
A into the tree from B. In the example, the `this.element` property of
`ComponentA` may be mutated "from the outside".

So, `data-nanomorph-component-id` allows encoding the inverse of
`isSameNode`: if the old and the new node have different component IDs,
they will not be morphed, but replaced.